Converting means and standard deviations

Converting estimated means and standard deviations in experimental and contorol groups to the effect size estimates and the within studies standard errors vector

Usage

convert_mean(n1, m1, s1, n2, m2, s2, pooled = FALSE)

Arguments

n1

A vector of number of observations in experimental group

m1

A vector of estimated mean in experimental group

s1

A vector of standard deviation in experimental group

n2

A vector of number of observations in experimental group

m2

A vector of estimated mean in experimental group

s2

A vector of standard deviation in experimental group

pooled

logical; if TRUE, a pooled variance is used. The default is FALSE.

Value

A data.frame of study data.

y: A numeric vector of the effect size estimates.

se: A numeric vector of the within studies standard error estimates.

Examples

require("flexmeta")
#> Loading required package: flexmeta
#> Warning: there is no package called 'flexmeta'
data("clbp")
dat <- convert_mean(clbp$n1, clbp$m1, clbp$s1, clbp$n2, clbp$m2, clbp$s2)
print(dat)
#>         y       se
#> 1    2.00 3.144921
#> 2    3.00 3.262923
#> 3   -2.00 4.132601
#> 4    3.30 3.745650
#> 5    5.00 3.745650
#> 6    8.30 5.884007
#> 7    1.00 8.179348
#> 8   -6.00 2.805140
#> 9    3.50 4.104256
#> 10   0.50 4.081157
#> 11  -5.84 5.313543
#> 12  -3.98 5.279749
#> 13   4.50 3.489772
#> 14  -2.00 2.066928
#> 15  -4.00 2.073189
#> 16 -37.60 2.829912
#> 17   4.00 7.205716
#> 18 -19.40 1.593289
#> 19  -8.00 2.120946
#> 20   0.06 1.891965
#> 21  -0.20 2.664159
#> 22  10.80 7.260702
#> 23 -27.20 8.747905